Added voice handling to server.

+++ VoiceCraft.Mobile/VoiceCraft.Mobile/Network/VoiceCraftParticipant.cs
@@ -88,7 +88,7 @@ namespace VoiceCraft.Mobile.Network
             this.PacketCount = PacketCount;
         }
 
-        public void SetVolume(int Volume)
+        public void SetVolume(float Volume)
         {
             FloatProvider.Volume = Volume;
         }
